chore: organize files into components folder
