+++ apps/site/src/app/(main)/portal/@applicant/components/timeline/StatusImage.tsx
